refactor and add to README
