+++ wikiconv/corpus-uploader/README.md
@@ -10,6 +10,12 @@ https://console.cloud.google.com/storage/browser/wikidetox-wikiconv-public-datas
 
 ## Figshare
 
+The `./ts-bin/figshare-uploader.ts` file is a typescript nodejs tool that uses
+the [Figshare API](https://docs.figshare.com/) to upload files to Figshare
+articles from [Google Cloud Storage](https://cloud.google.com/storage/).
+
+Usage:
+
 1. Make a copy of the token template file, and place it into a local `config`
    directory:
 
@@ -27,12 +33,24 @@ rsync --ignore_existing config.template.ts config/config.ts
 
 1. Run `ts-node ./ts-bin/figshare-uploader.ts`
 
-Note: for large uploads this can be slow; approx a week for the English corpus.
+The script will tell you about any missing files (in cloud storage but not
+having an entry in the Figshare article). The script skips all existing
+files/parts that are uploaded, and only uploads new parts that were missing. It
+takes a parameter `--concurrently=N` where `N` is a number for the concurrent
+uploads to use (default is 5). It is safe to cancel the script and restart it.
 
-IN PROGRESS: this is a rewrite in nodejs so that we can *easily* support concurrent uploads,
-restarting, command line flags for language, and stream directly from google
-cloud file to Figshare. This should speed things up considerably, estimated
-probably 20x speedup.
+TODO: make the script create new articles/files for files in cloud storage that
+are needed. For the wikiconv corpus, we created the files using the
+`old/upload.sh` script, but it fails to actually upload the parts.
+
+### Description of Figshare things here
+
+* `ts-lib/figshare_types.ts` contains common types for the JSON objects returned
+  by the Figshare API.
+* `ts-lib/cloud_storage_util.ts` provides a couple of helper functions for
+  working with Google Cloud storage.
+* `ts-lib/util.ts` provides some helper funcions (some set operations, sleep,
+  and pretty printing)
 
 ## Kaggle
