--- wikiconv/corpus-uploader/requirements.txt
@@ -1,3 +1,0 @@
-requests
-google-cloud-storage
-absl-py
