+++ wikiconv/corpus-uploader/ts-lib/figshare_types.ts
