+++ wikiconv/corpus-uploader/tsconfig.json
@@ -26,12 +26,13 @@
     // "noUnusedParameters": true,
     "removeComments": true,
     "strictNullChecks": true,
-    // Turned on in call (because it can't be turned off otherwise).
+    // Turned on by shell argument (because it can't be turned off otherwise).
     // "sourceMap": true,
     "target": "es6",
     "outDir": "build/js-bin/"
   },
   "include": [
-    "./ts-bin/**/*.ts"
+    "./ts-bin/**/*.ts",
+    "./ts-lib/**/*.ts"
   ]
 }
